splits up 3box components adds txbchain id to connectable networks adds the txbchain genesis config for reference
