+++ contracts/txbchain.spec.json
@@ -1,0 +1,42 @@
+{
+  "config": {
+    "chainId": 6874585,
+    "homesteadBlock": 0,
+    "eip150Block": 0,
+    "eip150Hash": "0x0000000000000000000000000000000000000000000000000000000000000000",
+    "eip155Block": 0,
+    "eip158Block": 0,
+    "byzantiumBlock": 0,
+    "constantinopleBlock": 0,
+    "petersburgBlock": 0,
+    "istanbulBlock": 0,
+    "clique": {
+      "period": 15,
+      "epoch": 15000
+    }
+  },
+  "nonce": "0x0",
+  "timestamp": "0x5f7c5cca",
+  "extraData": "0x000000000000000000000000000000000000000000000000000000000000000027a6013158094898ea4fe9b4c1a312b2d87987410000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000",
+  "gasLimit": "0x7270e0",
+  "difficulty": "0x1",
+  "mixHash": "0x0000000000000000000000000000000000000000000000000000000000000000",
+  "coinbase": "0x0000000000000000000000000000000000000000",
+  "alloc": {
+    "27a6013158094898ea4fe9b4c1a312b2d8798741": {
+      "balance": "0x200000000000000000000000000000000000000000000000000000000000000"
+    },
+    "f54a9a6e2A740c4B22F3fe1375F5e2bBa6660f7C": {
+      "balance": "0x56bc75e2d63100000"
+    },
+    "E231B4e55fE1D0Afb3e746e64E78eEffB5b599d1": {
+      "balance": "0x56bc75e2d63100000"
+    },
+    "85670f01716376BFB751822C4474fd0989779A96": {
+      "balance": "0x56bc75e2d63100000"
+    }
+  },
+  "number": "0x0",
+  "gasUsed": "0x0",
+  "parentHash": "0x0000000000000000000000000000000000000000000000000000000000000000"
+}
