+++ src/modules/connectors.ts
@@ -1,5 +1,5 @@
-import { InjectedConnector } from '@web3-react/injected-connector'
+import { InjectedConnector } from '@web3-react/injected-connector';
 
 export const injected = new InjectedConnector({
-     supportedChainIds: [1, 3, 4, 5, 17, 42, 1337, 5778] 
-})
+  supportedChainIds: [1, 3, 4, 5, 17, 42, 1337, 5778, 6874585],
+});
